+++ autumn/src/extract.rs
@@ -649,6 +649,7 @@ mod tests {
             "/blobs",
             Duration::from_secs(3600),
             SigningKey::random(),
+            vec![],
         )
         .unwrap();
 
@@ -689,6 +690,7 @@ mod tests {
             "/blobs",
             Duration::from_secs(3600),
             SigningKey::random(),
+            vec![],
         )
         .unwrap();